Refactoring Ramsey experiments into CYCLIC

+++ src/qibocal/update.py
@@ -31,12 +31,12 @@ def readout_frequency(freq: float, platform: Platform, qubit: QubitId):
 
 def bare_resonator_frequency(freq: float, platform: Platform, qubit: QubitId):
     """Update rbare frequency value in platform for specific qubit."""
-    platform.calibration.single_qubits[qubit].resonator.bare_frequency = int(freq)
+    platform.calibration.single_qubits[qubit].resonator.bare_frequency = freq
 
 
 def dressed_resonator_frequency(freq: float, platform: Platform, qubit: QubitId):
     """Update rbare frequency value in platform for specific qubit."""
-    platform.calibration.single_qubits[qubit].resonator.dressed_frequency = int(freq)
+    platform.calibration.single_qubits[qubit].resonator.dressed_frequency = freq
 
 
 def readout_amplitude(amp: float, platform: Platform, qubit: QubitId):
@@ -72,11 +72,11 @@ def drive_duration(
         duration = duration[0]
     if rx90:
         platform.update(
-            {f"native_gates.single_qubit.{qubit}.RX90.0.1.duration": int(duration)}
+            {f"native_gates.single_qubit.{qubit}.RX90.0.1.duration": duration}
         )
     else:
         platform.update(
-            {f"native_gates.single_qubit.{qubit}.RX.0.1.duration": int(duration)}
+            {f"native_gates.single_qubit.{qubit}.RX.0.1.duration": duration}
         )
 
 
@@ -138,7 +138,7 @@ def mean_exc_states(excited_state: list, platform: Platform, qubit: QubitId):
 
 def readout_fidelity(fidelity: float, platform: Platform, qubit: QubitId):
     """Update fidelity of single shot classification."""
-    platform.calibration.single_qubits[qubit].readout.fidelity = float(fidelity)
+    platform.calibration.single_qubits[qubit].readout.fidelity = fidelity
 
 
 def virtual_phases(
@@ -158,7 +158,7 @@ def virtual_phases(
     )
 
 
-def CZ_duration(duration: int, platform: Platform, pair: QubitPairId):
+def CZ_duration(duration: float, platform: Platform, pair: QubitPairId):
     """Update CZ duration for specific pair."""
     platform.update(
         {f"native_gates.two_qubit.{f'{pair[0]}-{pair[1]}'}.CZ.0.1.duration": duration}
@@ -172,7 +172,7 @@ def CZ_amplitude(amp: float, platform: Platform, pair: QubitPairId):
     )
 
 
-def iSWAP_duration(duration: int, platform: Platform, pair: QubitPairId):
+def iSWAP_duration(duration: float, platform: Platform, pair: QubitPairId):
     """Update iSWAP_duration duration for specific pair."""
     platform.update(
         {f"native_gates.two_qubit.{f'{pair[0]}-{pair[1]}'}.CZ.0.1.duration": duration}
@@ -186,12 +186,12 @@ def iSWAP_amplitude(amp: float, platform: Platform, pair: QubitPairId):
     )
 
 
-def t1(t1: int, platform: Platform, qubit: QubitId):
+def t1(t1: float, platform: Platform, qubit: QubitId):
     """Update t1 value in platform for specific qubit."""
     platform.calibration.single_qubits[qubit].t1 = tuple(t1)
 
 
-def t2(t2: int, platform: Platform, qubit: QubitId):
+def t2(t2: float, platform: Platform, qubit: QubitId):
     """Update t2 value in platform for specific qubit."""
     platform.calibration.single_qubits[qubit].t2 = tuple(t2)
 
@@ -213,7 +213,7 @@ def drag_pulse_beta(beta: float, platform: Platform, qubit: QubitId):
 
 def sweetspot(sweetspot: float, platform: Platform, qubit: QubitId):
     """Update sweetspot parameter in platform for specific qubit."""
-    platform.calibration.single_qubits[qubit].qubit.sweetspot = float(sweetspot)
+    platform.calibration.single_qubits[qubit].qubit.sweetspot = sweetspot
 
 
 def flux_coefficients(
@@ -221,7 +221,7 @@ def flux_coefficients(
 ):
     """Update flux-amplitude relation parameters for specific qubit."""
     platform.calibration.single_qubits[qubit].qubit.flux_coefficients = [
-        float(value) for value in flux_coefficients
+        value for value in flux_coefficients
     ]
 
 
@@ -230,10 +230,10 @@ def flux_offset(offset: float, platform: Platform, qubit: QubitId):
     platform.update({f"configs.{platform.qubits[qubit].flux}.offset": offset})
 
 
-def frequency_12_transition(frequency: int, platform: Platform, qubit: QubitId):
+def frequency_12_transition(frequency: float, platform: Platform, qubit: QubitId):
     channel = platform.qubits[qubit].drive_extra[1, 2]
     platform.update({f"configs.{channel}.frequency": frequency})
-    platform.calibration.single_qubits[qubit].qubit.frequency_12 = int(frequency)
+    platform.calibration.single_qubits[qubit].qubit.frequency_12 = frequency
 
 
 def drive_12_amplitude(amplitude: float, platform: Platform, qubit: QubitId):
@@ -244,13 +244,11 @@ def drive_12_amplitude(amplitude: float, platform: Platform, qubit: QubitId):
 
 def drive_12_duration(duration: int | tuple | list, platform: Platform, qubit: QubitId):
     """Update drive duration value in platform for specific qubit."""
-    platform.update(
-        {f"native_gates.single_qubit.{qubit}.RX12.0.1.duration": int(duration)}
-    )
+    platform.update({f"native_gates.single_qubit.{qubit}.RX12.0.1.duration": duration})
 
 
 def coupling(g: float, platform: Platform, qubit: QubitId):
-    platform.calibration.single_qubits[qubit].readout.coupling = float(g)
+    platform.calibration.single_qubits[qubit].readout.coupling = g
 
 
 def kernel(kernel: np.ndarray, platform: Platform, qubit: QubitId):
